refactor: reorganize project into domain, application, infrastructure, and presentation layers with
