--- src/application/port/MonitoringService.ts
@@ -1,10 +1,0 @@
-import { SmartFurnitureHookupID } from "@domain/SmartFurnitureHookupID";
-
-export interface MonitoringService {
-  registerSmartFurnitureHookup(
-    id: SmartFurnitureHookupID,
-    endpoint: string,
-  ): Promise<void>;
-
-  disconnectSmartFurnitureHookup(endpoint: string): Promise<void>;
-}
