+++ src/domain/errors.ts
