+++ src/infrastructure/HTTPMonitoringService.ts
@@ -1,8 +1,8 @@
-import { MonitoringService } from "@application/port/MonitoringService";
+import { MonitoringService } from "@application/outbound/MonitoringService";
 import { SmartFurnitureHookupID } from "@domain/SmartFurnitureHookupID";
 import axios from "axios";
 
-export class MonitoringServiceImpl implements MonitoringService {
+export class HTTPMonitoringService implements MonitoringService {
   private readonly MONITORING_SERVICE_URI =
     process.env.MONITORING_SERVICE_URI ||
     `http://${process.env.MONITORING_SERVICE_HOST || "monitoring"}:${process.env.MONITORING_SERVICE_PORT || 3003}`;
