+++ src/infrastructure/mongo/MongooseSmartFurnitureHookupRepository.ts
@@ -1,4 +1,3 @@
-import { SmartFurnitureHookupRepository } from "@domain/ports/SmartFurnitureHookupRepository";
 import { SmartFurnitureHookup } from "@domain/SmartFurnitureHookup";
 import { SmartFurnitureHookupID } from "@domain/SmartFurnitureHookupID";
 import { v4 as uuidv4, validate } from "uuid";
@@ -9,8 +8,9 @@ import {
   SmartFurnitureHookupEndpointConflictError,
   SmartFurnitureHookupNameConflictError,
   SmartFurnitureHookupNotFoundError,
-} from "@domain/errors/errors";
+} from "@domain/errors";
 import { MongoServerError } from "mongodb";
+import { SmartFurnitureHookupRepository } from "@application/outbound/SmartFurnitureHookupRepository";
 
 export class MongooseSmartFurnitureHookupRepository implements SmartFurnitureHookupRepository {
   private handleMongoSmartFurnitureHookupConflict(
