+++ src/infrastructure/mongo/mongoose/SmartFurnitureHookupDocument.ts
