+++ src/infrastructure/mongo/mongoose/SmartFurnitureHookupModel.ts
