--- src/interfaces/web-api/dependencies.ts
@@ -1,30 +1,0 @@
-import { router } from "./routes/router";
-
-import { SmartFurnitureHookupController } from "./controllers/SmartFurnitureHookupController";
-import { SmartFurnitureHookupServiceImpl } from "@application/SmartFurnitureHookupServiceImpl";
-import { AuthMiddleware } from "./middlewares/AuthMiddleware";
-import { MongooseSmartFurnitureHookupRepository } from "@storage/mongo/MongooseSmartFurnitureHookupRepository";
-import { MonitoringServiceImpl } from "@interfaces/MonitoringServiceImpl";
-
-// ===== Repository =====
-export const smartFurnitureHookupRepository =
-  new MongooseSmartFurnitureHookupRepository();
-
-// ===== Services =====
-export const monitoringService = new MonitoringServiceImpl();
-
-export const smartFurnitureHookupService = new SmartFurnitureHookupServiceImpl(
-  smartFurnitureHookupRepository,
-  monitoringService,
-);
-
-// ===== Controllers =====
-const smartFurnitureHookupController = new SmartFurnitureHookupController(
-  smartFurnitureHookupService,
-);
-
-// ===== Middlewares =====
-export const authMiddleware = new AuthMiddleware();
-
-// ===== Router =====
-export const apiRouter = router(smartFurnitureHookupController, authMiddleware);
