+++ src/presentation/web-api/middlewares/AuthMiddleware.ts
@@ -1,6 +1,6 @@
 import { type NextFunction, type Request, type Response } from "express";
 import axios from "axios";
-import { InvalidTokenError } from "@interfaces/web-api/middlewares/authMiddewareErrors";
+import { InvalidTokenError } from "@presentation/web-api/middlewares/authMiddewareErrors";
 
 export class AuthMiddleware {
   private readonly USER_SERVICE_URI =
