+++ src/presentation/web-api/routes/internal/internalRoutes.ts
@@ -1,6 +1,6 @@
 import { Router } from "express";
 import { smartFurnitureHookupRoutes } from "./smartFurnitureHookupRoutes";
-import { SmartFurnitureHookupController } from "@interfaces/web-api/controllers/SmartFurnitureHookupController";
+import { SmartFurnitureHookupController } from "@presentation/web-api/controllers/SmartFurnitureHookupController";
 
 export function internalRoutes(
   smartFurnitureHookupController: SmartFurnitureHookupController,
