+++ src/presentation/web-api/routes/smartFurnitureHookupRoutes.ts
