+++ src/test/interfaces/web-api/app.ts
@@ -1,7 +1,7 @@
 import "dotenv/config";
 import express from "express";
 import { apiRouter } from "./dependencies";
-import { errorHandler } from "@interfaces/web-api/middlewares/errorHandlerMiddleware";
+import { errorHandler } from "@presentation/web-api/middlewares/errorHandlerMiddleware";
 
 const app = express();
